feat: add core module and API
