+++ indiana_c/__init__.py
@@ -1,24 +1,12 @@
-from .generation import CORE_PROMPT, generate_text, reason_loop
-from .reflection import reflect
 from .model import IndianaC, IndianaCConfig
 from .monitor import SelfMonitor
 from .quantize import quantize_2bit
-from .logger import (
-    ThoughtComplexityLogger,
-    estimate_complexity_and_entropy,
-    thought_logger,
-)
+from .reflection import reflect
 
 __all__ = [
     "IndianaC",
     "IndianaCConfig",
-    "generate_text",
-    "reason_loop",
-    "reflect",
     "quantize_2bit",
     "SelfMonitor",
-    "CORE_PROMPT",
-    "ThoughtComplexityLogger",
-    "estimate_complexity_and_entropy",
-    "thought_logger",
+    "reflect",
 ]
