+++ requirements.txt
@@ -1,0 +1,5 @@
+fastapi
+uvicorn
+torch
+numpy
+tokenizers
