+++ tests/test_logger.py
@@ -1,4 +1,7 @@
-from indiana_c.logger import estimate_complexity_and_entropy, ThoughtComplexityLogger
+from indiana_core import (
+    estimate_complexity_and_entropy,
+    ThoughtComplexityLogger,
+)
 
 
 def test_estimate_complexity_and_entropy_keywords():
