+++ tests/test_reasoning.py
@@ -9,14 +9,18 @@ from pathlib import Path
 
 import torch
 
-from indiana_c.generation import generate_consistent_text, generate_with_think, reason_loop
-from indiana_c.tokenizer import tokenizer
+from indiana_core import (
+    generate_consistent_text,
+    generate_with_think,
+    reason_loop,
+    tokenizer,
+)
 
 
 def test_generate_with_think_returns_thought_and_final() -> None:
     """Ensure ``generate_with_think`` yields both text and metadata."""
 
-    with patch("indiana_c.generation.generate_text", return_value=("thought", {"c": 1})) as mock_gen:
+    with patch("indiana_core.generate_text", return_value=("thought", {"c": 1})) as mock_gen:
         result = generate_with_think("prompt")
 
     # The wrapper should request reasoning metadata and return the tuple as-is
@@ -30,11 +34,11 @@ def test_consistency_improves_with_multiple_attempts() -> None:
     side_effect = ["B", "A", "A"]
 
     # Single attempt may yield an inconsistent answer
-    with patch("indiana_c.generation.generate_with_think", side_effect=side_effect):
+    with patch("indiana_core.generate_with_think", side_effect=side_effect):
         single = generate_consistent_text("prompt", n=1)
 
     # Multiple attempts should recover the majority answer "A"
-    with patch("indiana_c.generation.generate_with_think", side_effect=side_effect):
+    with patch("indiana_core.generate_with_think", side_effect=side_effect):
         multi = generate_consistent_text("prompt", n=3)
 
     assert single != "A"
@@ -60,10 +64,10 @@ def test_reason_loop_alternates_and_logs() -> None:
             return torch.cat([idx, addition], dim=1)
 
     with (
-        patch("indiana_c.generation.IndianaC", DummyModel),
-        patch("indiana_c.generation.quantize_2bit", lambda _: None),
-        patch("indiana_c.generation.SelfMonitor.__init__", return_value=None),
-        patch("indiana_c.generation.SelfMonitor.log") as mock_log,
+        patch("indiana_core.IndianaC", DummyModel),
+        patch("indiana_core.quantize_2bit", lambda _: None),
+        patch("indiana_core.SelfMonitor.__init__", return_value=None),
+        patch("indiana_core.SelfMonitor.log") as mock_log,
     ):
         result = reason_loop("Q", max_steps=1)
